backend/modules/gcp/requirements.txt
Flask==3.1.1
google-cloud-kms==3.5.1
grpcio==1.64.1
requests==2.32.4
gunicorn==23.0.0